rename configuration to config
